# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRUD Laravel ",
        "description": "API para administrar multimples servicios y relaciones entre tablas",
        "contact": {
            "email": "jhonfredyaya04@gmail.com"
        },
        "license": {
            "name": "Repocitorio",
            "url": "https://github.com/JhonFac/Crud_Laravel.git"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/challenges/page/{id}": {
            "get": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Obtiene todos los challenges",
                "description": "Display a listing of the resource.",
                "operationId": "5380f426a793cb14275f7a9df7c4ed6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Numero de la pagina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de challenges"
                    }
                }
            }
        },
        "/api/challenges": {
            "post": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Crea un nuevo challenge",
                "description": "Store a newly created resource in storage.",
                "operationId": "63ffd18fb5f5371ec322250145cbb046",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "difficulty": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "challenge creado exitosamente"
                    }
                }
            }
        },
        "/api/challenges/{id}": {
            "get": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Obtiene el challenge por ID",
                "description": "Display the specified resource.",
                "operationId": "ed8928d8cd2720e0dcce6cd0e29767f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Numero de la pagina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de challenges"
                    }
                }
            },
            "put": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Actualiza un challenge existente",
                "description": "Update the specified resource in storage.",
                "operationId": "8f55380ce146081226d3559b05c43570",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del challenge a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "difficulty": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "challenge actualizado exitosamente"
                    },
                    "404": {
                        "description": "challenge no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Elimina un challenge existente",
                "description": "Remove the specified resource from storage.",
                "operationId": "8621e1614a29d01b58dd2c4384d0afba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del challenge a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "challenge eliminado exitosamente"
                    },
                    "404": {
                        "description": "challenge no encontrado"
                    }
                }
            }
        },
...
    },
...
}